requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
jinja2
python-multipart
openai
python-docx
